Post rebase tidy and move rule api functions to api folder
