+++ apps/rule-manager/client/src/ts/components/api/getRule.ts
