+++ apps/rule-manager/client/src/ts/components/hooks/useRules.ts
@@ -1,4 +1,4 @@
-import {useEffect, useState} from 'react';
+import { useEffect, useState } from 'react';
 
 export function useRules() {
     const { location } = window;
