+++ apps/rule-manager/conf/routes
@@ -19,6 +19,6 @@ POST    /rules                      controllers.RulesController.create()
 +nocsrf
 GET     /rules/:id                  controllers.RulesController.rule(id: Int)
 +nocsrf
-
+POST    /rules/:id                  controllers.RulesController.update(id: Int)
 # Map static resources from the /public folder to the correct URL path
-GET     /*file                controllers.Assets.versioned(path="/public/", file: Asset)
+GET     /*file                      controllers.Assets.versioned(path="/public/", file: Asset)
